Unit testing for player class and more refinement of player class
