+++ cards/GameView.resx
